Improved tool setup + removed stupid comments
